requirements.txt
asgiref==3.7.2
certifi==2023.7.22
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
Faker==19.2.0
mysqlclient==2.1.1
Pillow==10.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
